+++ skosprovider/skos.py
@@ -25,6 +25,9 @@ class Label:
     '''
     A :term:`SKOS` Label.
     '''
+    
+    uri = None
+    '''A :term:`URI` for this label.'''
 
     label = None
     '''
@@ -51,18 +54,27 @@ class Label:
     The valid types for a label
     '''
 
-    def __init__(self, label, type="prefLabel", language="und"):
+    def __init__(self, label, type="prefLabel", language="und", uri=None):
         self.label = label
         self.type = type
         if not language:
             language = 'und'
+        if uri and not is_uri(uri):
+            raise ValueError('%s is not a valid URI.' % uri)
+        self.uri = uri
         if tags.check(language):
             self.language = language
         else:
             raise ValueError('%s is not a valid IANA language tag.' % language)
 
     def __eq__(self, other):
-        return self.__dict__ == (other if type(other) == dict else other.__dict__)
+        if type(other) == dict:
+            if self.uri:
+                return self.uri == other['uri']
+            return self.label == other['label'] and self.type == other['type'] and self.language == other['language'] 
+        if self.uri:
+            return self.uri == other.uri
+        return self.label == other.label and self.type == other.type and self.language == other.language 
 
     def __ne__(self, other):
         return not self == other
@@ -75,6 +87,9 @@ class Label:
         :param string type: The type to be checked.
         '''
         return type in Label.valid_types
+
+    def is_xl(self):
+        return self.uri is not None
 
     def __repr__(self):
         return "Label('{}', '{}', '{}')".format(self.label, self.type, self.language)
@@ -590,8 +605,8 @@ def filter_labels_by_language(labels, language, broader=False):
 
 def dict_to_label(dict):
     '''
-    Transform a dict with keys `label`, `type` and `language` into a
-    :class:`Label`.
+    Transform a dict with keys `label`, `type`, `language` and `uri`
+    into a :class:`Label`.
 
     Only the `label` key is mandatory. If `type` is not present, it will
     default to `prefLabel`. If `language` is not present, it will default
@@ -604,7 +619,8 @@ def dict_to_label(dict):
         return Label(
             dict['label'],
             dict.get('type', 'prefLabel'),
-            dict.get('language', 'und')
+            dict.get('language', 'und'),
+            uri = dict.get('uri', None)
         )
     except (KeyError, AttributeError, TypeError):
         return dict